- Move groups related actions to DeviceManager
- Removed cache updates from getXxxxById
